requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine learning
scikit-learn>=0.24.0

# Geospatial
# pykrige>=1.6.0  # Optional: Only needed if using Universal Kriging

# Parallel processing
joblib>=1.0.0
tqdm>=4.60.0

# Configuration
pyyaml>=5.4.0

# Jupyter for reproduction notebook
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Optional: For PDF reading (if analyzing papers)
# PyPDF2>=1.26.0

# Development dependencies (optional)
# pytest>=6.2.0
# pytest-cov>=2.12.0
# black>=21.5b0
# flake8>=3.9.0
